Ubuntu desktop does not come with aptitude or cURL
Since Ubuntu 10.10, aptitude has been left out of the default Ubuntu
desktop install. This commit changes the linux and linux-prerequisites
scripts to check if aptitude is installed: if it is the script continues
as normal; if it is not apt-get is used to install aptitude, and then
the script continues as normal.

This also updates README to use wget instead of cURL, and installs cURL
later.