Switch Thin to Unicorn
As the Rails application scales to more Heroku dynos, a lack of
concurrency can result in [poor performance][1] and an over-provisioning
of resources. Modifying the app to handle requests more efficiently is
a simple process with immediate benefits.

Unicorn is a [concurrent web server][2] that spawns several processes
within a single dyno without requiring concurrency or threading
awareness in your app.  Applications that migrate to Unicorn often
require fewer dynos and see increased performance.

[1]: https://blog.heroku.com/archives/2013/2/16/routing_performance_update
[2]: https://blog.heroku.com/archives/2013/2/27/unicorn_rails