Use Homebrew for `hub` on OS X
The previous method places the `hub` command in ~/bin/hub, which shows
up in the dotfiles repo for thoughtbot/dotfiles users. We could
gitignore the `bin/hub` gem there, but the hub README recommends using
Homebrew or a package manager when it is available.

This leaves the Linux version functionally the same.